Improve DPI awareness
- add a few DPI checks in key places
- add app manifest, setting "System (Enhanced)" (a.k.a. GDI Scaling) mode by default

I've also tested the application-managed Per-monitor-aware mode.
If desired, the user can override to this via the Compatibility tab in the EXE's
Properties in Windows Explorer. This mode means that menus and dialog boxes will be fully scaled
(layout will be a bit disrupted around fixed-size icons), and the app will be fully responsible
for DPI scaling. You still only get 1x, 2x and 4x modes as has always been the case, and 1x mode
may be a bit broken because of the doubled menu bar for large scale factors. Output will be
pixel-crisp, so some users may prefer this, but it's not the default because it's a bit of a
surprise and maybe not perfectly implemented.